+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Chotsiri 2019 lumefantrine ([doi:10.1002/cpt.1531](https://doi.org/10.1002/cpt.1531)) -- children aged 6-59 months with uncomplicated malaria, with and without severe acute malnutrition.
+
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.
 
 - Add Decrane 2023 oxyfluorfen ([doi:10.1016/j.crtox.2023.100138](https://doi.org/10.1016/j.crtox.2023.100138)) -- rat (Sprague-Dawley) and extrapolated human.